requirements.txt
crewai
liteLLM
edgartools
requests
pandas
pydantic
python-dotenv
plotly
PyYAML
beautifulsoup4
lxml
openai
